Conflicts:
	PMP2.2/src/main/resources/templates/assetstype_manage.html
	PMP2.2/src/main/resources/templates/firmware.html
	PMP2.2/src/main/resources/templates/index.html
	PMP2.2/src/main/resources/templates/typeRequest.html
	PMP2.2/src/main/resources/templates/users.html
	PMP2.2/src/main/resources/templates/vendor.html

+++ PMP2.2/src/main/java/com/example/demo/repository/DataCenterRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.repository.query.Param;
 
 import com.example.demo.entities.DataCenter;
 
-public interface DataCenterRepository  extends  JpaRepository<DataCenter, Long>{
+public interface DataCenterRepository  extends  JpaRepository<DataCenter, Integer>{
 	
 	public Page<DataCenter> findByAdress(String n, Pageable pageable);
