+++ PMP2.2/src/main/java/com/example/demo/repository/TypeExpensesRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.entities.TypeExpenses;
 
-public interface TypeExpensesRepository extends JpaRepository<TypeExpenses, Long> {
+public interface TypeExpensesRepository extends JpaRepository<TypeExpenses, Integer> {
 	
 
 }
